move migrations to migrations{,_test}.go
