+++ overlord/snapstate/snapstate.go
@@ -464,60 +464,3 @@ func GadgetInfo(s *state.State) (*snap.Info, error) {
 
 	return nil, state.ErrNoState
 }
-
-// MigrateToTypeInState implements a state migration to have the snap type in the snap state of each setup snap. To be used in overlord/migrations.go.
-func MigrateToTypeInState(s *state.State) error {
-	var stateMap map[string]*SnapState
-
-	err := s.Get("snaps", &stateMap)
-	if err == state.ErrNoState {
-		// nothing to do
-		return nil
-	}
-	if err != nil {
-		return err
-	}
-
-	for snapName, snapState := range stateMap {
-		if snapState.CurrentSideInfo() == nil {
-			continue
-		}
-		typ := snap.TypeApp
-		snapInfo, err := readInfo(snapName, snapState.CurrentSideInfo())
-		if err != nil {
-			logger.Noticef("Recording type for snap %q: cannot retrieve info, assuming it's a app: %v", snapName, err)
-		} else {
-			logger.Noticef("Recording type for snap %q: setting to %q", snapName, snapInfo.Type)
-			typ = snapInfo.Type
-		}
-		snapState.SetType(typ)
-	}
-
-	s.Set("snaps", stateMap)
-	return nil
-}
-
-// MigrateToCurrentRevision implements a state migration to have the snap Current revision in the snap state of each setup snap. Used in overlord/migrations.go.
-func MigrateToCurrentRevision(s *state.State) error {
-	var stateMap map[string]*SnapState
-
-	err := s.Get("snaps", &stateMap)
-	if err == state.ErrNoState {
-		// nothing to do
-		return nil
-	}
-	if err != nil {
-		return err
-	}
-
-	for _, snapState := range stateMap {
-		n := len(snapState.Sequence)
-		if n == 0 {
-			continue
-		}
-		snapState.Current = snapState.Sequence[n-1].Revision
-	}
-
-	s.Set("snaps", stateMap)
-	return nil
-}
